Move FileDistributionFactory to a more suitable package
